backend/requirements.txt
flask
flask-cors
pymupdf
langchain
langchain-community
langchain-openai
openai
python-dotenv
tiktoken
flask-sqlalchemy
flask-migrate
sendgrid
pdfplumber
